Separate container validation logic in a different class
