+++ core/REST/REST_Controller.php
@@ -2,6 +2,7 @@
 namespace Carbon_Fields\REST;
 
 use Carbon_Fields\Container\Container;
+use Carbon_Fields\REST\Container_Validator;
 
 class REST_Controller {
 	
@@ -17,14 +18,6 @@ class REST_Controller {
 	 */
 	protected $vendor = 'carbon-fields';
 
-	/**
-	 * The containers holding the fields
-	 * for the response
-	 * 
-	 * @var array
-	 */
-	public $containers = [];
-
 	protected $special_field_types = [
 		'complex',
 		'relationship',
@@ -35,6 +28,20 @@ class REST_Controller {
 		'html',
 		'separator',
 	];
+
+	/**
+	 * The containers holding the fields
+	 * for the response
+	 * 
+	 * @var array
+	 */
+	public $containers = [];
+
+	public $validator;
+
+	public function __construct() {
+		$this->validator = new Container_Validator();
+	}	
 
 	public function get_data( $type, $id  = '') {
 		$response = [];
@@ -65,7 +72,7 @@ class REST_Controller {
 
 	public function filter_containers( $type, $id = '' ) {
 		return array_filter( Container::$active_containers, function( $container ) use ( $type, $id ) {
-			return $this->is_valid_container( $container, $type, $id );
+			return $this->validator->is_valid_container( $container, $type, $id );
 		} );
 	}
 
@@ -85,54 +92,6 @@ class REST_Controller {
 		return $this->vendor;
 	}
 
-	public function is_valid_container( $container, $type, $id ) {
-		if ( $container->type !== $type ) {
-			return false;
-		}
-
-		$type_to_lower = strtolower( $type );
-
-		return call_user_func( [ $this, "is_valid_{$type_to_lower}_container"], $container, $id );
-	}
-
-	public function is_valid_theme_options_container( $container, $id ) {
-		return true;
-	}
-
-	public function is_valid_post_meta_container( $container, $id ) {
-		return $container->is_valid_save_conditions( $id );
-	}
-
-	public function is_valid_user_meta_container( $container, $id ) {
-		return $this->is_valid_post_meta_container( $container, $id );
-	}
-
-	public function is_valid_term_meta_container( $container, $id ) {
-		$term = get_term( $id );
-
-		if ( empty( $term ) || is_wp_error( $term ) ) { 
-			return false;
-		}
-		
-		$taxonomy = $term->taxonomy;
-
-		if ( ! in_array( $taxonomy, $container->settings['taxonomy'] ) ) {
-			return false;
-		}
-
-		if ( $container->settings['show_on_level'] ) { 
-			
-			$show_level = $container->settings['show_on_level'];
-			$term_level = self::get_term_level( $term );
-
-			if ( $term_level !== $show_level ) {
-				return false;
-			}
-		}
-
-		return true;
-	}
-
 	public function load_generic_field_value( $field ) {
 		return $field->get_value();
 	}
@@ -147,15 +106,5 @@ class REST_Controller {
 
 	public function load_association_field_value( $field ) {
 		return $this->load_relationship_field_value( $field );
-	}
-
-	public static function get_term_level( $term ) {
-		$ancestors = [];	
-		while ( ! is_wp_error( $term ) && ! empty( $term->parent ) && ! in_array( $term->parent, $ancestors ) ) {
-			$ancestors[] = intval( $term->parent );
-			$term        = get_term( $term->parent );
-		}
-
-		return count( $ancestors ) + 1;
 	}
 }
